Charge conjugation boundary conditions for gauge fields implemented as a policy class, changing the nature of covariant Cshifts used in plaquettes, rectangles and staples.
As a result same code is used for the plaq and rect action independent of the BC type.

Should probably isolate the BC in a separate class that Gimpl takes as a template param.
Do the same with smearing policies.

This would then allow composition of BC with smearing etc....